% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_other_filetypes.R
\name{flagstats_to_total_primary_reads}
\alias{flagstats_to_total_primary_reads}
\title{Extract total primary alignments from flagstats file}
\usage{
flagstats_to_total_primary_reads(flagstats)
}
\arguments{
\item{flagstats}{path to a 'flagstats' file created by \code{samtools flagstat <bam>}}
}
\value{
total number of primary alignments (whether or not they pass or fail QC)
}
\description{
Since most aligners keep unmapped reads and label them as primary -
this value is usually a good proxy
for the total number of reads fed into the aligner.
}
\details{
Note both forward and reverse reads contribute independently to this count.
}
