% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomics.R
\name{parse_kraken_report}
\alias{parse_kraken_report}
\title{Parse a Kraken2 report}
\usage{
parse_kraken_report(path)
}
\arguments{
\item{path}{Path to a tabular KrakenUniq report file.}
}
\value{
A \code{data.table} with standardised column names.
}
\description{
Reads a Kraken2 report and standardises key column names so output includes the following:
\code{taxid}, \code{fullname}, \code{reads_covered_by_clade}, \code{reads_assigned_directly}.
}
\examples{
path = system.file("reports/kraken2.report.tsv")
parse_kraken_report(path)
}
