% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomics.R
\name{parse_krakenuniq_report}
\alias{parse_krakenuniq_report}
\title{Parse a KrakenUniq report}
\usage{
parse_krakenuniq_report(path)
}
\arguments{
\item{path}{Path to a tabular KrakenUniq report file.}
}
\value{
A \code{data.table} with standardised column names.
}
\description{
Reads a KrakenUniq report and standardises key column names so output includes the following:
\code{taxid}, \code{fullname}, \code{reads_covered_by_clade}, \code{reads_assigned_directly}.
}
\examples{
path = system.file("reports/krakenuniq.report.tsv")
parse_krakenuniq_report(path)
}
