% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_delete_taxid_and_children}
\alias{taxonomy_delete_taxid_and_children}
\title{Delete a taxid and all descendants}
\usage{
taxonomy_delete_taxid_and_children(taxonomy, taxid)
}
\arguments{
\item{taxonomy}{taxonomy (igraph) object produced by \code{\link[=parse_taxonomy]{parse_taxonomy()}}}

\item{taxid}{a single taxid to delete.}
}
\value{
A taxonomy (igraph) with \code{taxid} and all descendants removed.
}
\description{
Deletes a taxid and all of its children from a taxonomy.
}
\examples{
path_to_taxdb = system.file(package="taxgraph", "taxonomies/taxDB")
taxonomy = parse_taxonomy(path_to_taxdb, type = "taxdb")
taxonomy_delete_taxid_and_children(taxonomy, taxid = "10376")
}
