% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_filter_for_descendants}
\alias{taxonomy_filter_for_descendants}
\title{Filter for a taxid and its descendants}
\usage{
taxonomy_filter_for_descendants(taxonomy, taxid)
}
\arguments{
\item{taxonomy}{taxonomy (igraph) object produced by \code{\link[=parse_taxonomy]{parse_taxonomy()}}}

\item{taxid}{a single taxid to retain along with its descendants.}
}
\value{
A subgraph (igraph) containing \code{taxid} and all descendant nodes.
}
\description{
Returns a subgraph containing a taxid and all of its descendants.
}
\examples{
path_to_taxdb = system.file(package="taxgraph", "taxonomies/taxDB")
taxonomy = parse_taxonomy(path_to_taxdb, type = "taxdb")
taxonomy_filter_for_descendants(taxonomy, taxid = "10376")
}
