% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_filter_for_ranks}
\alias{taxonomy_filter_for_ranks}
\title{Filter for specific ranks}
\usage{
taxonomy_filter_for_ranks(taxonomy, ranks)
}
\arguments{
\item{taxonomy}{taxonomy (igraph) object produced by \code{\link[=parse_taxonomy]{parse_taxonomy()}}}

\item{ranks}{a vector of ranks to filter for.}
}
\value{
A subgraph (igraph) of taxonomy including only specified \code{ranks}.
}
\description{
Filters a taxonomy for nodes with rank in \code{ranks}, while preserving
connectivity where possible.
}
\examples{
path_to_taxdb = system.file(package="taxgraph", "taxonomies/taxDB")
taxonomy = parse_taxonomy(path_to_taxdb, type = "taxdb")
taxonomy_filter_for_ranks(taxonomy, ranks = c("family", "genus", "species"))
}
