% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_filter_for_taxids}
\alias{taxonomy_filter_for_taxids}
\title{Filter for specific taxids}
\usage{
taxonomy_filter_for_taxids(taxonomy, taxids)
}
\arguments{
\item{taxonomy}{taxonomy (igraph) object produced by \code{\link[=parse_taxonomy]{parse_taxonomy()}}}

\item{taxids}{a vector of taxids to filter for.}
}
\value{
A subgraph (igraph) of taxonomy including only specified \code{taxids}.
}
\description{
Filters a taxonomy for specific taxids while preserving connectivity where possible.
For example if you have graph \strong{A -> B -> C -> D} and filter for \strong{B} & \strong{D}
all other vertices are removed by vertex suppression, producing the graph:
\strong{B -> D}
}
