% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_remove_taxids_with_low_read_support}
\alias{taxonomy_remove_taxids_with_low_read_support}
\title{Remove taxids with low read support}
\usage{
taxonomy_remove_taxids_with_low_read_support(
  taxonomy,
  min_reads_covered_by_clade = 1
)
}
\arguments{
\item{taxonomy}{taxonomy (igraph) object produced by \code{\link[=parse_taxonomy]{parse_taxonomy()}}
and annotated with \code{reads_covered_by_clade}.}

\item{min_reads_covered_by_clade}{minimum reads required to retain a taxid.}
}
\value{
A subgraph (igraph) of taxonomy including only supported taxids.
}
\description{
Filters a taxonomy to retain only nodes with \code{reads_covered_by_clade}
greater than or equal to \code{min_reads_covered_by_clade}, while preserving
connectivity where possible.
}
\examples{
path_to_taxdb = system.file(package="taxgraph", "taxonomies/taxDB")
taxonomy = parse_taxonomy(path_to_taxdb, type = "taxdb")
# assuming taxonomy has been annotated with read counts
taxonomy_remove_taxids_with_low_read_support(taxonomy, min_reads_covered_by_clade = 10)
}
