% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{parse_taxonomy}
\alias{parse_taxonomy}
\title{Parse taxonomy}
\usage{
parse_taxonomy(path, type = c("taxdb", "ktaxonomy", "ncbi"))
}
\arguments{
\item{path}{path to a taxonomy file}

\item{type}{type of taxonomy file described by path (taxDB, ktaxonomy, ncbi)}
}
\description{
Parse taxonomy
}
