% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_igraph_tree}
\alias{plot_igraph_tree}
\title{Plot a quick igraph tree}
\usage{
plot_igraph_tree(taxonomy, circular = FALSE)
}
\arguments{
\item{taxonomy}{taxonomy object created from \code{\link[=parse_taxonomy]{parse_taxonomy()}} or}

\item{circular}{}
}
\description{
Plot a quick igraph tree
}
