% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sunburst_plotly}
\alias{plot_sunburst_plotly}
\title{Plot a sunburst chart from a taxonomy graph}
\usage{
plot_sunburst_plotly(taxonomy, force = FALSE)
}
\arguments{
\item{taxonomy}{An igraph object with a \code{reads_covered_by_clade} vertex attribute, e.g. produced by \code{\link[=report_to_graph]{report_to_graph()}}.}

\item{force}{Logical; set to TRUE to allow plotting large graphs.}
}
\value{
A plotly sunburst plot.
}
\description{
Plot a sunburst chart from a taxonomy graph
}
\examples{
path_to_taxdb = system.file(package="taxgraph", "taxonomies/taxDB")
path_to_report = system.file(package= "taxgraph", "reports/krakenuniq.report.tsv")

taxonomy = parse_taxonomy(path_to_taxdb, type = "taxdb")
report_dataframe = parse_krakenuniq_report(path_to_report)

report_graph = report_to_graph(report_dataframe, taxonomy)
plot_sunburst_plotly(report_graph)
}
